backend/requirements.txt
numpy==1.26.4
pandas==2.2.1
Flask==2.0.1
Flask-CORS==3.0.10
requests>=2.31.0
python-dotenv==0.19.0
Werkzeug==2.0.1
